put repo editing dialog in its own component for reuse: RepoEditor.vue
